Updates for planned Openwire Support
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@473801 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/ActiveMQMapMessage.h
@@ -330,7 +330,7 @@ namespace commands{
          * @param value - the Long value to set in the Map
          * @throws CMSException
          */
-        virtual void setLong( const std::string& name, long value );
+        virtual void setLong( const std::string& name, long long value );
 
         /**
          * Returns the Short value of the Specified name
